fixing fmt cuda error

+++ src/umpire/alloc/MallocAllocator.hpp
@@ -82,7 +82,7 @@ struct MallocAllocator {
       UMPIRE_ERROR(runtime_error,
                    fmt::format("cudaDeviceGetAttribute(pageableMem = {}, cudaDevAttrPageableMemoryAccess = {}, cdev = "
                                "{}) failed with error: {}",
-                               pageableMem, cudaDevAttrPageableMemoryAccess, cdev, cudaGetErrorString(err)));
+                               pageableMem, static_cast<int>(cudaDevAttrPageableMemoryAccess), cdev, cudaGetErrorString(err)));
     }
     if (pageableMem)
       return true;
