+++ src/umpire/resource/FileMemoryResource.cpp
@@ -159,7 +159,7 @@ bool FileMemoryResource::isPageable() noexcept
     UMPIRE_ERROR(runtime_error,
                  fmt::format("cudaDeviceGetAttribute(pageableMem = {}, cudaDevAttrPageableMemoryAccess = {}, cdev = "
                              "{}) failed with error: {}",
-                             pageableMem, cudaDevAttrPageableMemoryAccess, cdev, cudaGetErrorString(err)));
+                             pageableMem, static_cast<int>(cudaDevAttrPageableMemoryAccess), cdev, cudaGetErrorString(err)));
   }
   if (pageableMem)
     return true;
